Add versioned API commands as a markdown doc
Adds a `-version` flag to `cmd/generator` which is used to pre-process
the `apidocs.json` before any fo the Go code is generated from it. This
flag and variable has two main purposes:
    * Generating a version markdown definition of all API commands.
    * Generating a `pkg/api/version.go` file which matches ECE_VERSION.

Additionally by introducing a makefile variable `ECE_VERSION`, where the
value is controlled from a single place even though it's used for
multiple purposes.

Furthermore, it adds a bit of documentation on how to update the swagger
definition and the steps to follow.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>