deps: Update swagger to v0.22.0 (#59)
Updates the swagger generator to the latest available version. This
causes a few things to change for the autogenerated client and models,
but none of them are breaking.

An important change to mention is that any mismatches between the JSON
response (i.e. statuscode) and the actual models which are casted into
will result in a panic where before an error was returned. This can have
unintended consequences for consuming clients.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>