plan: Fix bug where last step error was missed
Fixes a bug in the plan tracker logic which made it return the first
step which finished with status "error" rather than the last step. This
has many downstream effects but more notably, makes vacates miss the end
result of the vacate and potentially set a cluster vacate as "Sucessful"
when it is not.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>