multierror: Add package with Prefixed error type
Adds a new package under `pkg/multierror` which is aimed to be used
instead of `multierror.Error`. The main motivation behind this change is
to enforce a prefix to be used in all of the multierrors which will
provide more context to the error consumers.

It features multierror unpacking to keep all the errors same level and
in case the wrapped errors are also a Prefixed type, a prefix will be
added to each of the single errors to add that additional context to it.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>