plan: Add change tracker backed by Deployments API (#66)
This patch adds a new `TrackChange` function which uses the deployments
API to track any pending changes of a Deployment. It supports tracking
changes by specifying a `DeploymentID` or `ResourceID` and `Kind` which
is very similar to what the legacy plan tracker did.

This tracker, will track any pending plans of all the resources of the
specified Deployment unless `IgnoreDownstream` is set to `true`. In
which case maps the behavior of the legacy plan tracker.

Additionally, a `planutil` package aggregates both TrackChange and
Stream into a single function call with configurable stream output
formats.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>
Co-authored-by: Karen Cárcamo <karencfv@users.noreply.github.com>