cmdutil: Add DecodeFlags to reduce parsing code
Adds a new `DecodeFlags` function which through reflection and
`mapstructure` is able to decode a cobra.Command defined flags into a
structure's fields which match the flag's type and mapstructure tag.

By using this function we can greatly reduce the flag-parsing code and
offload most if not all of the flag parsing by using this function.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>