api: Move LoginUser from ecctl codebase
This patch contains a few changes aimed to making the SDK friendlier for
consumers, moving logic from external calls to implicit calls on the
API constructing function, controling the behavior through a bool flag
which still allows control over it when desired.

There's multierror changes which have been done to move over from the
multierror package owned from Hashicorp to our own. This wasn't entirely
needed for this patch but it makes the errors and wrapping of errors
that much nicer that it adds good value.

Also Moves the `UnwrapError` function from the `api` package to an
`apierror` subpackage and renaming the function to `Unwrap` this is due
to an import cyclic issue where `api` imports the `auth` package and in
order to be able to unwrap the errors returned in the `auth` package
`auth` would be importing `api` only so it can call `api.UnwrapError`.

Furthermore, now `auth` does unwrap the errors resulting in less cryptic
errors returned when the user cannot be logged in or the user's bearer
token cannot be refreshed.

Last, adds a `NewAuthWriter` function in the `auth` package so that an
`auth.Writer` can created by either using a populated `Username` and
`Password` or an `APIKey`. This also reduces the need to do that in all
of the multiple clients which need to instantiate an `auth.Writer` to
pass it to the `API` structure.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>