Further changes to multierror apierror and mock
First, fixes the linters.

Refactors the apierror.Unwrap to several private functions to split the
concerns and make the code more readable. After this patch, the function
will unpack any errors which are of type `*models.BasicFailedReply` to
be wrapped by a `multierror.Prefixed` rather than printing the Marshaled
JSON.

Adds a conditional to the `multierror.Prefixed` which ensures that the
appended error is `*multierror.Prefixed` and the Prefix matches both the
parent and the error which is being added, then the prefix is removed on
appending the error.

Now the `mock` package contains both sample `mock.Response` errors for
consumers to use conveniently as well as what the expected `error` type
for the added functions for testing purposes.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>