api: Add better API defaults and helpers
This changes the constructor behavior of `NewAPI` and the `Config`
validation now only accepts an `auth.APIKey` or `*auth.APIKey` as the
only valid AuthWriter's for the API when the host is set to the ESS host
(`https://api.elastic-cloud.com`). Also adds a constant: `ESSEndpoint`.

Additionally, adds a new function to the `deploymentapi` package:
`deploymentapi.RequestID`, which will always return a request ID to be
used in `deploymentapi.CreateParams`. It will either return an auto-
generated string of 64 characters or the string argument which is passed
to the function.

Last, it adds a new package called `deploymentsize` and a new function:
`Parse` which will parse a string in the format of `<size><unit`, valid
examples of parseable sizes are:

    * Gigabyte: `10g`, `0.5G`, `16gb`.
    * Megabyte: `1024m`, `2048mb`, `512m`.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>