mock: Add ability to assert received *http.Request
This patch adds the ability to assert certain *http.Request fields which
are sent to the http.RoundTripper on any of the request / reply flows
which are tested and asserted in our unit tests.

Adds the ability to assert what's sent over to the API so our unit tests
become more complete. Since it's an optional field, it doesn't enforce
that field to be set in the `mock.Response`, but rather make it optional
for those tests where it makes sense to do so.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>