api: Add WithRegion context for region bound calls
Adds a `context.Context` constructor (`api.WithRegion`) which creates a
new context with a saved value equal to the passed region. This is meant
to be used in any calls to the auto-generated API which are region-bound
(any `/platform` APIs).

When using `*api.CloudClientRuntime`, the path interpolation will happen
without having to explicitly handle it.

This change allows the `*api.API` object to be used accross multiple
regions rather than requiring a single instance of that structure per
region.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>