platformapi: Add region parameter to all APIs
This patch modifies all the platformapi public functions to have a
Region string required parameter which is then used to create a new
context.Context: api.WithRegion(context.Background(), params.Region)
and pass it as the API operation context.

Unit tests have been modified to include a request
assertion which ensures that the region path interpolation is used when
the /platform APIs are being called.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>