api: Add global route mappings and require region
Removes the deprecated `api.Config.Region` property and modifies `api.
CloudClientRuntime` to require a context.Context with a set region value
for the paths which require a region to be set or return an error.

`var globalPath map[string]bool` is introduced to allow the runtime to
be able to know when to error when the region path needs to be used and
a context with a region value set is missing.

Additionally, removes a few deprecated fields, variables properties and
the legacy plan tracker which used the clusters API.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>