configurationtemplateapi: Add Format field (#162)
This patch adds a new Format field to the GetTemplateParams, ListTemplatesParams and PullToFolderParams structs.

This will allow the use of the WithFormat() method, in order to specify which inofrmation is populated in the response.
If "cluster" is specified, cluster_template is pupulated in the response. If deployment is specified, deployment_template 
is populated in the response.