deploymentapi: Support List and filter templates
Adds a workaround for a missing API so that the deployment templates can
be obtained on ESS by using the `deploymenttemplateapi.List()` method
with a filter function that returns an error when the template is not
found.

The behavior is controlled with a boolean `AsList` in `GetParams` and
a bool named `DeploymentTemplateAsList` on the `NewPayloadParams`.

This boolean should be removed as soon as the GET endpoint for a single
deployment template is available on ESS.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>