sync/pool: Add FailFast parameter, stopping on err
This patch adds a new `FailFast` parameter to `pool.Params` which allows
consumers to toggle a new pool behavior where the pool will gracefully
stop when one of the workers returns with an error. This will stop
processing any additional work but try to continue processing the work
that's in-flight or cancel the execution when the `params.Timeout.Stop`
`time.Duration` is exceeded.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>