actions: Add scheduled api-validation job (#217)
Adds a new GH actions workflow to run the api-validation tests once
a day.

Additionally, I have removed the api-validation tests on each
make swagger as now we don't necessarily have the apidocs
on par with what's on production. We may have some endpoints
that are not publicly available yet for development purposes, so
there's no point in having these as they are likely to fail on occasion.