api: Add CustomTransport unifying RoundTripper(s)
Adds a new implementation of `http.RoundTripper` which aims to unify all
existing segregated RoundTrippers which have been marked as deprecated.

Additionally, adding support to configurable retries on request timeout
through two new `api.Config` fields:

* `Retries` and `RetryBackoff`.

To keep backwards compatibility, the retries are opt-in even though they
would improve the resilience of the SDK. They can always be made opt-out
in the future if we deem it necessary.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>