allocatorapi: Fix incorrect json error formatting
This patch fixes the returned error format when an allocatorapi.Vacate
returns with an error.

The problem wasn't on how the errors are marshalled when OutputFormat,
but on the `multierror.Prefixed` error and the unpacking that it does.
As it turns out, because the prefix from the error from `plan.Stream()`
isn't equal to the the multi error where it is flattened to through the
`merr.Append()` (`unpackErrors()`), as it was being squashed by calling
`fmt.Sprint(prefixed.Prefix, ": ", errElement.Error()),`. Already, this
isn't ideal and probably we might want to look more closely at using the
error wrapping `fmt.Errorf("%s: %w")`, and its implications.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>