apierror: Add new Error type with Unwrap methods
Adds a new `apierror.Error` type which is able to return the wrapped
underlying `error` that's returned from the `go-openapi/runtime` SDK.

This change is required to be able to assert the type of error returned
by an API call which relies on `go-openapi/runtime` SDK by type since
the previous `apierror.Unwrap` fell short in a number of places and
mutated the error itself returning either a `multierror.Prefixed` or
an `errors.errorString`, resulting in non-assertable errors for clients.

Now, upon calling `apierror.Unwrap` an `apierror.Error` type is returned
which implements the `errors.Unwrap` interface, allowing consumers to
perform assertions based on the error type with `errors.As` or `Unwrap`.

This is particularly helpful for the Terraform provider work that we're
doing, but will also prove useful for other work.

The output of `apierror.Error.Error()` remains unchanged, making this
change for all intents and purposes a non-breaking change when it comes
to the returned signature types and complying with the interfaces.

Two helpers have been added in `runtime.go`, which allow consumers to
assert certain status codes when the error is not mapped in the swagger
spec yet the API returns that status code.

Moreover, a constructing `apierror.Wrap()` function has been written to
help deprecate the previous `apierror.Unwrap()` and reflect more closely
what the code is doing, which is wrapping errors.

Last, all error assertions have been changed to check the error string
instead of the error type, for two reasons; First, it would be a waste
of engineering type to compose all those error assertions in the tests,
and second, the Go linting complains when an error is asserted via the
`reflect.DeepEqual()` method, calling for string comparisons rather than
the DeepEqual comparisons.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>