deploymentapi: Add OverrideCreateOrUpdateRequest
Adds a new public function `OverrideCreateOrUpdateRequest` which allows
setting certain overrides on the Deployment requests in the Create and
Update operations.

The function was private before making it part of the Create and Update
calls. This wasn't always desirable specially when pairing it with the
`--generate-payload` flag in `ecctl` where it wouldn't return the fields
that were overridden.

The function has been expanded to:

- Set overrides to all the application `ref_id` fields.
- ElasticsearchRefID overrides.
- Plugin overrides.
- Selection between node_roles and node_types when both are set.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>