release: Add branch GitHub action
Adds a GitHub action which creates a new branch when a tag is pushed
with the tag's MAJOR.MINOR. If the branch already exists, the action
will fail, but it can be ignored. GitHub actions has no way to set the
step as successful on failure, we'd need to modify the source code of
the used action.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>