release: Automate version updates, branching (#333)
This patch introduces a few changes, to automate the version updates and
branching when a GitHub tag is pushed to the repository. The additions
take the previous "branching" action and split it into two:
    * release
    * bump-main-version

`release` will create a GitHub release with the Changelog notes as the
body, bump the patch version, create a branch using the `MAJOR.MINOR`,
and commits the version update.

`bump-main-version` checks out the `master` branch and updates the minor
version. Since Major version updates are rare enough where they can be
done manually (still leveraging `make major`).

The version bumping leverages a pre-existing project which works pretty
well.

Signed-off-by: Marc Lopez <marc5.12@outlook.com>