Keep all Elasticsearch topologies when building update request.
The response from the API will already contain all topologies, but they were removed when building the update request. No we simply keep all elements as the API returns them.

This is mainly needed when autoscaling is enabled. In that case the update request must contain all existing topology elements.