Keep all Elasticsearch topology elements when building update request. (#416)
This change keeps all topology elements as they are returned by the API. This is mainly needed when autoscaling is enabled. In that case the update request must contain all existing topology elements even if they are size=0.

The code used to remove all elements with size=0, but it doesn't hurt to have them so they are now included with every request.