Fix vacate for integrations-server kind.
When vacating an integrations-server instance, the automatic tracking of the request currently did not work. This is due to the API lacking support for integrations-server (APM and integrations-server are returned as APM).

The underlying problem is that searching for the resource by using the resource-id + kind (APM) will not find anything.

Since the resource-ids are unique, we can just search by resource-id to find the relevant resource to track.