Fix vacate for integrations-server kind. (#455)
When vacating an integrations-server instance, the automatic tracking of the request currently did not work. This is due to the API lacking support for integrations-server (APM and integrations-server are returned as kind APM). Effectively this means 'kind' returned by this API can't be used as it could be wrong.

This change replaces usages of `kind` by `resource-id`. As the `resource-id` is a unique identifier across all resources, additionally filtering by `kind` is not necessary.