Change deployment-alias type to *string. (#472)
The deployment-alias field has the following behavior:
- Not added in request -> Don't update field
- Set to empty value -> Remove deployment alias
- Set to non-empty value -> Set alias to that value

So with the old spec, it wasn't possible to remove the alias after it was set, as due to omitempty an empty value is not added to the request. This is now possible.

The get-deployment has a slightly different behavior: If there is no deployment-alias, the field is not set, if there is a deployment-alias, the field is set. So for the get and responses, it is better to keep the type as it is.